requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.36
pyodbc==5.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
